# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/superadmin/admins": {
      "get": {
        "operationId": "SuperadminController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "List all admin/superadmin users (superadmin only)",
        "tags": [
          "Superadmin — Team"
        ]
      },
      "post": {
        "operationId": "SuperadminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Create a new admin/superadmin user (superadmin only)",
        "tags": [
          "Superadmin — Team"
        ]
      }
    },
    "/superadmin/admins/{id}": {
      "delete": {
        "description": "Hard-deletes the user; user_roles cascades. A superadmin may not delete their own account.",
        "operationId": "SuperadminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Delete an admin/superadmin user (superadmin only)",
        "tags": [
          "Superadmin — Team"
        ]
      }
    },
    "/superadmin/admins/{id}/password": {
      "put": {
        "operationId": "SuperadminController_setPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAdminPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Directly set an admin/superadmin user's password (superadmin only)",
        "tags": [
          "Superadmin — Team"
        ]
      }
    }
  },
  "info": {
    "title": "MedAlign Backend API",
    "description": "MedAlign backend API. Authentication uses an HttpOnly `access_token` cookie set on POST /auth/2fa/verify. All non-GET requests require an `x-csrf-token` header matching the value returned by the CSRF middleware (CSRF is skipped when NODE_ENV=development).\n\n## Local test seed data (run `pnpm seed:pj` against your local DB)\n\n**Clinic portal logins** (password `SeedClinic1!` for both):\n- `clinic-vitality@medalign-seed.example.com` — clinic `vitality-hormone-nyc`\n- `clinic-apex@medalign-seed.example.com` — clinic `apex-peptide-telehealth`\n\n**Superadmin login** (for `/admin/clinics*` and `/admin/patients*`): `superadmin@medalign-seed.example.com` / `SeedAdmin1!`\n\n**Soft-deleted patient** (locked out of sign-in; use it to exercise the admin patient deleted-state and lockout paths): `patient-deleted@medalign-seed.example.com` / `SeedPatient1!`\n\n**2FA:** in `NODE_ENV=development` the 6-digit code is printed to the server logs (LoggingEmailSender) instead of being emailed. Read it from the console and POST it to `/auth/2fa/verify` to obtain the session cookie.\n\nFor patient-flow testing, sign up a new user via POST /auth/signup, then submit an assessment (POST /assessments) to get a `sessionId` + `claimToken`.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
...
      "CreateAdminDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe",
            "description": "Admin's full name"
          },
          "email": {
            "type": "string",
            "example": "jane@medalign.example.com",
            "description": "Email address"
          },
          "password": {
            "type": "string",
            "example": "NewPassw0rd!",
            "description": "Password (min 8 characters)",
            "minLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "superadmin"
            ],
            "example": "admin"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role"
        ]
      },
      "SetAdminPasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string",
            "example": "NewPassw0rd!",
            "minLength": 8
          }
        },
        "required": [
          "newPassword"
        ]
      }
    }
  }
}